+++ Controllers/InvitationController.cs
@@ -35,7 +35,7 @@ namespace wie_doet_de_afwas.Controllers
                 return UnauthorizedJson();
             }
 
-            float averageScore = 0;
+            double averageScore = 0;
             foreach (var gm in group.GroupMembers)
             {
                 averageScore += gm.Score;
